feat(gcm): Add GCM::Encryptor::encrypt_and_authenticate
This function is exposed to the GCM consumer because it takes plaintext
and associated data and directly returns ciphertext together with the
auth tag.
Note that normally we would want to return a single
std::vector<std::uint8_t> with "ciphertext || auth_tag" to ensure the
user always transmit both parts. However for the assignment it is useful
to receive them in two seperate vectors.